+++ frontend/src/Pages/Explore/Explore.tsx
@@ -1,0 +1,5 @@
+const ExplorePage = () => {
+  return <>Explore Page</>;
+};
+
+export default ExplorePage;
