+++ frontend/src/api/constants.ts
@@ -1,0 +1,7 @@
+const API_ROOT = "http://localhost:8000/api/v1";
+
+export const loginURL = `${API_ROOT}/users/create-session`;
+export const signupURL = `${API_ROOT}/users/signup`;
+export const createJobURL = `${API_ROOT}/users/createjob`;
+
+// export const signUpURL = `${API_ROOT}/users/signup`;
