--- frontend/src/core/components/inputLabel.tsx
@@ -1,8 +1,0 @@
-export function InputLabel(props: any) {
-  const { title } = props;
-  return (
-    <>
-      <h2 className="text-xl mx-1">{title}</h2>
-    </>
-  );
-}
