--- frontend/src/core/constants.ts
@@ -1,5 +1,0 @@
-const API_ROOT = "http://localhost:8000/api/v1";
-
-export const loginURL = `${API_ROOT}/users/create-session`;
-export const signUpURL = `${API_ROOT}/users/signup`;
-export const createJobURL = `${API_ROOT}/users/createjob`;
