--- frontend/src/feature/jobs-board/jobs-board.tsx
@@ -1,7 +1,0 @@
-export function JobsBoard() {
-  return (
-    <>
-      <h1>hello</h1>
-    </>
-  );
-}
