--- frontend/src/feature/user-auth/components/registrationPage.tsx
@@ -1,187 +1,0 @@
-import { useNavigate } from "react-router-dom";
-import { signup } from "../../../core/auth";
-import { useForm } from "react-hook-form";
-import { useState } from "react";
-import {
-  Button,
-  Stack,
-  TextField,
-  Select,
-  MenuItem,
-  SelectChangeEvent,
-  InputLabel,
-  FormControl,
-} from "@mui/material";
-
-type FormValues = {
-  name: string;
-  email: string;
-  password: string;
-  confirmPassword: string;
-};
-
-export function RegistrationPage() {
-  const navigate = useNavigate();
-  const [role, setRole] = useState("Manager");
-
-  const form = useForm<FormValues>({
-    defaultValues: { name: "", email: "", password: "", confirmPassword: "" },
-  });
-
-  const { register, handleSubmit, formState, watch } = form;
-  const { errors } = formState;
-
-  const onSubmit = (data: FormValues) => {
-    console.log("form submitted");
-    console.log(data);
-    signup(
-      data.email,
-      data.password,
-      data.confirmPassword,
-      data.name,
-      role,
-      navigate
-    );
-  };
-
-  return (
-    <>
-      <div className="mx-auto bg-slate-50 content flex flex-col justify-center items-center">
-        <div className=" p-4  border rounded bg-white">
-          <div className="text-3xl justify-center text-black mb-2 ">
-            Create New Account
-          </div>
-          <form onSubmit={handleSubmit(onSubmit)} noValidate>
-            <Stack spacing={2} width={400}>
-              <TextField
-                label="Name"
-                type="text"
-                {...register("name", {
-                  required: "Name is required",
-                })}
-                error={!!errors.name}
-                helperText={errors.name?.message}
-                sx={{
-                  "& label": { paddingLeft: (theme) => theme.spacing(1) },
-                  "& input": { paddingLeft: (theme) => theme.spacing(2.5) },
-                  "& fieldset": {
-                    paddingLeft: (theme) => theme.spacing(1.5),
-                    borderRadius: "10px",
-                  },
-                }}
-              />
-
-              <TextField
-                label="Email Id"
-                type="email"
-                {...register("email", {
-                  required: "Email is required",
-                  pattern: {
-                    value: /^[\w-\.]+@([\w-]+\.)+[\w-]{2,4}$/,
-                    message: "Enter a valid email",
-                  },
-                })}
-                error={!!errors.email}
-                helperText={errors.email?.message}
-                sx={{
-                  "& label": { paddingLeft: (theme) => theme.spacing(1) },
-                  "& input": { paddingLeft: (theme) => theme.spacing(2.5) },
-                  "& fieldset": {
-                    paddingLeft: (theme) => theme.spacing(1.5),
-                    borderRadius: "10px",
-                  },
-                }}
-              />
-              <TextField
-                label="Password"
-                type="password"
-                {...register("password", {
-                  required: "Password is required",
-                })}
-                error={!!errors.password}
-                helperText={errors.password?.message}
-                sx={{
-                  "& label": {
-                    paddingLeft: (theme) => theme.spacing(1),
-                  },
-                  "& input": { paddingLeft: (theme) => theme.spacing(2.5) },
-                  "& fieldset": {
-                    paddingLeft: (theme) => theme.spacing(1.5),
-                    borderRadius: "10px",
-                  },
-                }}
-              />
-              <TextField
-                label="Confirm password"
-                type="password"
-                {...register("confirmPassword", {
-                  required: "Password is required",
-                  validate: (val: string) => {
-                    if (watch("password") !== val) {
-                      return "Passwords don't match";
-                    }
-                  },
-                })}
-                error={!!errors.confirmPassword}
-                helperText={errors.confirmPassword?.message}
-                sx={{
-                  "& label": {
-                    paddingLeft: (theme) => theme.spacing(1),
-                  },
-                  "& input": { paddingLeft: (theme) => theme.spacing(2.5) },
-                  "& fieldset": {
-                    paddingLeft: (theme) => theme.spacing(1.5),
-                    borderRadius: "10px",
-                  },
-                }}
-              />
-              <FormControl>
-                <InputLabel id="role-id">Role</InputLabel>
-                <Select
-                  value={role}
-                  labelId="role-id"
-                  label="Role"
-                  id="role"
-                  onChange={(e: SelectChangeEvent) => {
-                    setRole(e.target.value);
-                  }}
-                >
-                  <MenuItem value={"Manager"}>Manager</MenuItem>
-                  <MenuItem value={"Applicant"}>Applicant</MenuItem>
-                </Select>
-              </FormControl>
-              <Button
-                type="submit"
-                variant="contained"
-                color="primary"
-                style={{
-                  background: "#FF5353",
-                  borderRadius: "10px",
-                  textTransform: "none",
-                  fontSize: "16px",
-                }}
-              >
-                Sign up
-              </Button>
-            </Stack>
-          </form>
-          <div className="mx-auto"></div>
-          <br />
-          <div className="mv-1 border-t mx-16" />
-          <div className="flex justify-center">
-            <p className="-mt-3 bg-white px-3 text-[#CCCCCC]">OR</p>
-          </div>
-          <br />
-          <p
-            className="text-[#656565] text-center"
-            onClick={() => {
-              navigate("/login");
-            }}
-          >
-            Already have an Account? Login Here
-          </p>
-        </div>
-      </div>
-    </>
-  );
-}
