--- frontend/src/feature/user-auth/constants/authConstants.ts
@@ -1,4 +1,0 @@
-const API_ROOT = "http://localhost:8000/api/v1";
-
-const loginURL = `${API_ROOT}/users/create-session`;
-const signupURL = `${API_ROOT}/users/signup`;
